refactor: move component sandbox methods to specific folder
