+++ src/renderer/services/plugin-manager/component/hooks.js
@@ -1,0 +1,9 @@
+export const hooks = [
+  'created',
+  'beforeMount',
+  'mounted',
+  'beforeUpdate',
+  'updated',
+  'beforeDestroy',
+  'destroyed'
+]
