+++ src/renderer/services/plugin-manager/plugin-component-sandbox.js
@@ -1,170 +1,8 @@
-import { PLUGINS } from '@config'
 import path from 'path'
-
-const hooks = [
-  'created',
-  'beforeMount',
-  'mounted',
-  'beforeUpdate',
-  'updated',
-  'beforeDestroy',
-  'destroyed'
-]
-
-export function prepareContext (vueContext, component) {
-  const context = vueContext._data || {}
-
-  const keys = ['$nextTick', '$refs', '_c', '_v', '_s', '_e', '_m', '_l', '_u']
-  for (let key of keys) {
-    let vueContextItem = vueContext[key]
-
-    if (key === '$refs' && vueContextItem) {
-      key = 'refs'
-      vueContextItem = {}
-      const badGetters = [
-        'attributes',
-        'children',
-        'childNodes',
-        'contentDocument',
-        'contentWindow',
-        'firstChild',
-        'firstElementChild',
-        'lastChild',
-        'lastElementChild',
-        'nextElementSibling',
-        'nextSibling',
-        'offsetParent',
-        'ownerDocument',
-        'parentElement',
-        'parentNode',
-        'shadowRoot',
-        'previousElementSibling',
-        'previousSibling',
-        '$root',
-        '__vue__'
-      ]
-      const badSetters = [
-        'innerHTML',
-        'outerHTML'
-      ]
-      vueContext.$nextTick(() => {
-        for (const elKey in vueContext.$refs) {
-          const element = vueContext.$refs[elKey]
-
-          if (!element.tagName || element.tagName.toLowerCase() === 'iframe') {
-            continue
-          }
-
-          for (const badGetter of badGetters) {
-            element.__defineGetter__(badGetter, () => console.log('🚫'))
-          }
-
-          for (const badSetter of badSetters) {
-            element.__defineSetter__(badSetter, () => console.log('🚫'))
-          }
-
-          vueContextItem[elKey] = element
-        }
-      })
-    }
-
-    context[key] = vueContextItem
-  }
-
-  for (const computedName of Object.keys(component.computed || {})) {
-    context[computedName] = vueContext[computedName]
-  }
-
-  for (const methodName of Object.keys(component.methods || {})) {
-    context[methodName] = function () {
-      return component.methods[methodName].apply(prepareContext(vueContext, component))
-    }
-  }
-
-  return context
-}
-
-export function defineContext (componentName, proxyComponent, baseComponent, vue) {
-  for (const methodName of Object.keys(proxyComponent.methods || {})) {
-    proxyComponent.methods[methodName] = baseComponent.methods[methodName]
-  }
-
-  // Build Vue component
-  const vmComponent = vue.component(componentName, proxyComponent)
-
-  // Fix context of "data" method
-  if (baseComponent.data) {
-    vmComponent.options.data = function () { return baseComponent.data.apply(prepareContext(this, baseComponent)) }
-  }
-
-  // Fix context of "computed" methods - also removes global computed methods
-  for (const computedName of Object.keys(vmComponent.options.computed)) {
-    vmComponent.options.computed[computedName] = function () {}
-  }
-
-  vmComponent.options.created = [function () {
-    for (const computedName of Object.keys(this.$options.computed)) {
-      if (baseComponent.computed && baseComponent.computed[computedName]) {
-        this.$options.computed[computedName] = baseComponent.computed[computedName].bind(
-          prepareContext(this)
-        )
-        this._computedWatchers[computedName].getter = baseComponent.computed[computedName].bind(
-          prepareContext(this)
-        )
-
-        try {
-          this._computedWatchers[computedName].run()
-        } catch (error) {
-          console.error(error)
-        }
-      }
-
-      if (!baseComponent.computed || !baseComponent.computed[computedName]) {
-        delete this.$options.computed[computedName]
-
-        try {
-          this._computedWatchers[computedName].teardown()
-        } catch (error) {
-          console.error(error)
-        }
-
-        delete this._computedWatchers[computedName]
-
-        for (const watcherId in this._watchers) {
-          if (this._watchers[watcherId].getter.name === computedName) {
-            try {
-              this._watchers[watcherId].teardown()
-            } catch (error) {
-              console.error(error)
-            }
-
-            break
-          }
-        }
-      }
-    }
-
-    if (baseComponent.created) {
-      return baseComponent.created.apply(prepareContext(this))
-    }
-  }]
-
-  // Fix context of hooks
-  hooks
-    .filter(hook => Object.prototype.hasOwnProperty.call(baseComponent, hook))
-    .filter(hook => hook !== 'created')
-    .forEach(prop => {
-      const componentMethod = baseComponent[prop]
-      const hookMethod = function () { return componentMethod.apply(prepareContext(this, baseComponent)) }
-      if (Array.isArray(vmComponent.options[prop])) {
-        vmComponent.options[prop] = [hookMethod]
-      } else {
-        vmComponent.options[prop] = hookMethod
-      }
-    })
-
-  return vmComponent
-}
+import { PLUGINS } from '@config'
+import { prepareContext } from './component/prepare-context'
+import { defineContext } from './component/define-context'
+import { hooks } from './component/hooks'
 
 export class PluginComponentSandbox {
   constructor ({
@@ -210,7 +48,7 @@ export class PluginComponentSandbox {
   render () {
     const renderedTemplate = this.vm.run(
       `const Vue = require('vue/dist/vue.common.js')
-      const compiled = Vue.compile('${this.compiled.template}')
+      const compiled = Vue.compile(${JSON.stringify(this.compiled.template)})
       const prepareContext = ${prepareContext.toString()}
       const component = {}
       if (compiled.staticRenderFns.length) {
